Revert to old config
