+++ composer.json
@@ -20,9 +20,9 @@
         "ext-pdo": "*",
         "erusev/parsedown": "1.7.4",
         "fakerphp/faker": "^1.23.1",
-        "illuminate/console": "^8.0|^9.0|^10.0|^11.0",
-        "illuminate/routing": "^8.0|^9.0|^10.0|^11.0",
-        "illuminate/support": "^8.0|^9.0|^10.0|^11.0",
+        "illuminate/console": "^8.0|^9.0|^10.0|^11.0|^12.0",
+        "illuminate/routing": "^8.0|^9.0|^10.0|^11.0|^12.0",
+        "illuminate/support": "^8.0|^9.0|^10.0|^11.0|^12.0",
         "league/flysystem": "^1.1.4|^2.1.1|^3.0",
         "mpociot/reflection-docblock": "^1.0.1",
         "nikic/php-parser": "^5.0",
@@ -40,7 +40,7 @@
         "laravel/legacy-factories": "^1.3.0",
         "league/fractal": "^0.20",
         "nikic/fast-route": "^1.3",
-        "orchestra/testbench": "^6.0|^7.0|^8.0",
+        "orchestra/testbench": "^6.0|^7.0|^8.0|^v9.10.0",
         "pestphp/pest": "^1.21",
         "phpstan/phpstan": "^2.1.1",
         "phpunit/phpunit": "^9.0|^10.0",
@@ -54,7 +54,7 @@
             "Knuckles\\Camel\\": "camel/"
         },
         "files": [
-          "src/Config/strategies.php"
+          "src/Config/helpers.php"
         ]
     },
     "autoload-dev": {
