--- src/Config/Factory.php
@@ -1,11 +1,0 @@
-<?php
-
-namespace Knuckles\Scribe\Config;
-
-class Factory
-{
-    public static function make(Extracting $extracting, Output $output): array
-    {
-        return Serializer::toOldConfig($extracting, $output);
-    }
-}
