+++ tests/BaseLaravelTest.php
@@ -25,9 +25,9 @@ class BaseLaravelTest extends TestCase
         ]);
         $app['config']->set('database.default', 'sqlite');
         $app['config']->set('database.connections.sqlite', [
-            'driver'   => 'sqlite',
+            'driver' => 'sqlite',
             'database' => ':memory:',
-            'prefix'   => '',
+            'prefix' => '',
         ]);
         ScribeServiceProvider::$customTranslationLayerLoaded = false;
     }
@@ -49,59 +49,53 @@ class BaseLaravelTest extends TestCase
     {
         parent::setUp();
 
-        $this->setConfig(Config\Factory::make(
-            extracting: Config\Extracting::with(
-                routes: Config\Routes::match(prefixes: ['*']),
-                defaultGroup: 'Endpoints',
-                databaseConnectionsToTransact: [],
-                fakerSeedForExamples: 1234,
-                auth: Config\Extracting::auth(
-                    enabled: false,
-                    default: false,
-                    in: AuthIn::BEARER,
-                    useValue: env('SCRIBE_AUTH_KEY'),
-                    placeholder: '{YOUR_AUTH_KEY}',
-                ),
-                strategies: Config\Extracting::strategies(
-                    metadata: Config\Defaults::METADATA_STRATEGIES,
-                    urlParameters: Config\Defaults::URL_PARAMETERS_STRATEGIES,
-                    queryParameters: Config\Defaults::QUERY_PARAMETERS_STRATEGIES,
-                    headers: Config\Defaults::HEADERS_STRATEGIES,
-                    bodyParameters: Config\Defaults::BODY_PARAMETERS_STRATEGIES,
-                    responses: configureStrategy(
-                        Config\Defaults::RESPONSES_STRATEGIES,
-                        Strategies\Responses\ResponseCalls::withSettings(
-                            only: [],
-                            except: ['*'], // Disabled to speed up tests
-                            config: [
-                                'app.env' => 'documentation',
-                                // 'app.debug' => false,
-                            ],
-                            queryParams: [],
-                            bodyParams: [],
-                            fileParams: [],
-                            cookies: [],
-                        )),
-                    responseFields: Config\Defaults::RESPONSE_FIELDS_STRATEGIES,
-                )
-            ),
-            output: Config\Output::with(
-                type: Config\Output::laravelType(
-                    theme: Config\InHouseTheme::Legacy,
-                    docsUrl: '/apidocs'
-                ),
-                baseUrls: [
-                    "production" => config("app.base_url"),
+        $this->setConfig([
+            'title' => config('app.name').' API Documentation',
+            'type' => 'laravel',
+            'theme' => 'default',
+            'laravel.docs_url' => '/apidocs',
+            'base_url' => config("app.base_url"),
+            // Skip these for faster tests
+            'postman.enabled' => false,
+            'openApi.enabled' => false,
+            'routes' => [
+                [
+                    'match' => [
+                        'prefixes' => ['*'],
+                        'domains' => ['*'],
+                    ],
                 ],
-                // Skip these for faster tests
-                postman: Config\Output::postman(
-                    enabled: false,
+            ],
+            'groups.default' => 'Endpoints',
+            'auth' => [
+                'enabled' => false,
+                'default' => false,
+                'in' => AuthIn::BEARER->value,
+                'name' => 'key',
+                'use_value' => env('SCRIBE_AUTH_KEY'),
+                'placeholder' => '{YOUR_AUTH_KEY}',
+            ],
+            'strategies' => [
+                'metadata' => Config\Defaults::METADATA_STRATEGIES,
+                'urlParameters' => Config\Defaults::URL_PARAMETERS_STRATEGIES,
+                'queryParameters' => Config\Defaults::QUERY_PARAMETERS_STRATEGIES,
+                'headers' => Config\Defaults::HEADERS_STRATEGIES,
+                'bodyParameters' => Config\Defaults::BODY_PARAMETERS_STRATEGIES,
+                'responses' => configureStrategy(
+                    Config\Defaults::RESPONSES_STRATEGIES,
+                    Strategies\Responses\ResponseCalls::withSettings(
+                        only: [],
+                        except: ['*'], // Disabled to speed up tests
+                        config: [
+                            'app.env' => 'documentation',
+                        ],
+                    )
                 ),
-                openApi: Config\Output::openApi(
-                    enabled: false,
-                ),
-            )
-        ));
+                'responseFields' => Config\Defaults::RESPONSE_FIELDS_STRATEGIES,
+            ],
+            'examples.faker_seed' => 1234,
+            'database_connections_to_transact' => [],
+        ]);
     }
 
     protected function setConfig($configValues): void
